+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Adds Bullet install generator. ([@abhaynikam][])
 * Adds Audit gems(bundler-audit, ruby_audit) install generator. ([@abhaynikam][])
 * Adds Pry install generator. ([@abhaynikam][])
+* Adds Active Storage Google install generator. ([@abhaynikam][])
 
 ## 0.2.0 (September 16th, 2020)
 * Adds JQuery install generator. ([@abhaynikam][])